package/libcpprestsdk: add dependency on Boost.System
libcpprestsdk searches for the Boost.System module in its
CMakeLists. Hence it should be selected as a dependency.

This does not fix any build failure, as boost-system was implicitly
selected by one of the other boost-* options that this package
selects, but an upcoming commit is going to change how boost-system is
selected by other boost-* modules, making this preparation change
necessary.

Signed-off-by: Michael Nosthoff <buildroot@heine.tech>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>