configs/iot-gate-imx8_ebbr: new defconfig
Add a defconfig to build an AArch64 disk image with a U-Boot based
firmware implementing the subset of UEFI defined by EBBR[1], as well as
a Linux OS booting with UEFI, to run on the IOT-GATE-iMX8 from Compulab.

The generated firmware binary can also be used to install or run another
OS supporting the EBBR specification.

This configuration is based on the work of Paul Liu[2] and is not related
with the official BSP from Compulab.

This has been tested on actual hardware.

While at it, add this defconfig to myself in DEVELOPERS.

[1] https://github.com/ARM-software/ebbr
[2] https://git.linaro.org/plugins/gitiles/people/paul.liu/systemready/build-scripts

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>
Cc: Ying-Chun Liu (Paul Liu) <paul.liu@linaro.org>
Signed-off-by: Julien Olivain <ju.o@free.fr>