package/refpolicy: fix upstream URL in help text
The previous URL [1] currently redirects to [2], which as of this
commit is about two months old and contains a readme that just says
"PoC by Callgh0st". When the package was added with commit [3]
the help text URL matched REFPOLICY_SITE, so presumably the help was
forgotten when the site URL was updated in [4], and there has been
some kind of hostile takeover of the former upstream repository since.

[1] https://github.com/TresysTechnology/refpolicy
[2] https://github.com/Exact-Realty/refpolicy
[3] https://gitlab.com/buildroot.org/buildroot/-/commit/bb442abdf18d6a4673ffa1f19e2a62f36c5d3038
[4] https://gitlab.com/buildroot.org/buildroot/-/commit/cfa018d9f300b07c9c01a5b1aeca49ee177dd0dc

Signed-off-by: Fiona Klute <fiona.klute@gmx.de>
[Julien: add commit ID which forgot to update the package homepage]
Signed-off-by: Julien Olivain <ju.o@free.fr>