package/kexec: bump to version 2.0.32
For release announce, see [1].

This commit also adds the support for LoongArch64, which was added in
upstream commit [2], first included in version v2.0.26. Note that
Buildroot introduced LoongArch64 support in commit [3].

This commit also updates the COPYING license file hash, after a minor
update in [4].

This release was signed with a different pgp key, so this commit
updates the key ID in comment.

[1] https://lists.infradead.org/pipermail/kexec/2025-October/034019.html
[2] https://git.kernel.org/pub/scm/utils/kernel/kexec/kexec-tools.git/commit/?id=1c8bf2dc0127b06f97a7973488c7f8cf9a5f7c19
[3] https://gitlab.com/buildroot.org/buildroot/-/commit/54895db9c10a26283b25a393cb6c872d72c245d1
[4] https://git.kernel.org/pub/scm/utils/kernel/kexec/kexec-tools.git/commit/?id=591ebf60649de53dd7a9c916ef4e831626ab1810

Cc: Jiaxun Yang <jiaxun.yang@flygoat.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>
Reviewed-by: Jiaxun Yang <jiaxun.yang@flygoat.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>